+++ ChangeLog.md
@@ -1,5 +1,15 @@
 # BlogMore ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added a `with_backlinks` configuration option that enables a "References &
+  mentions" section on individual post pages. When turned on, BlogMore scans
+  every post's Markdown content for internal links to other posts and shows
+  a collated list of back-linking posts.
+  ([#404](https://github.com/davep/blogmore/pull/404))
+
 ## v2.15.0
 
 **Released: 2026-04-24**